parse_mapping takes process_fns variable
parse_mapping now takes a new value, process_fns. process_fns is a dictionary of functions to apply to metadata categories. The keys should be the column headings, and the values should be functions which take a single value. For example, if the values in a column called "taxonomy" should be split on semi-colons before being added as metadata, and all other columns should be left as-is, process_fns should be:
      {'taxonomy': lambda x: x.split(';')}

This is being added specifically for the taxonomy use-case, but could also be useful in the future for type-casting values, etc.