parse_mapping allows users to override column headers
Users can now optionally pass a header list to parse_mapping. This is useful if a header line is not provided with the metadata, if users want to rename the fields, or if users want to include only the first n-1 fields where n is the number of headers provided (and the 1 is the first column header - typically either "SampleID" or "ObservationID").