Clean up sparse backend type reporting
The discrepancy in sparse backend type reporting between CSMat and ScipySparseMat occurred because CSMat was an old-style class and ScipySparseMat is new-style (inherits from object). This caused the formatting of the class name to differ. CSMat is now a new-style class.

The new output for both backends looks like:

     SparseObj type: CSMat
     SparseObj type: ScipySparseMat

This is a bit cleaner/clearer than what we had before (e.g., biom.backends.csmat.CSMat), since users are really only concerned with the name of the backend. The backend name output by ``biom show-install-info`` is now consistent with the name users must add to their BIOM config file.

Fixes #231.